Moddified models

+++ app/models/reviews.py
@@ -1,6 +1,6 @@
 from datetime import date
 
-from sqlalchemy import Date, ForeignKey, Integer, String
+from sqlalchemy import Date, ForeignKey, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.database import Base
@@ -16,6 +16,6 @@ class Review(Base):
     user_id: Mapped[int] = mapped_column(
         Integer, ForeignKey("users.id"), nullable=False
     )
-    rating: Mapped[int] = mapped_column(Integer, nullable=False)
+    rating: Mapped[float] = mapped_column(Numeric(2, 1), nullable=False)
     comment: Mapped[str] = mapped_column(String, nullable=False)
     created_at: Mapped[Date] = mapped_column(Date, default=date.today)
